Add files via upload
The topic of this project is SUPER MARKET BILLING MANAGEMENT SYSTEM. In this project, there are 3 floors having 3 stores each. Each store has its own protected database. The owner of that store has to enter the data like the name of the product, price and quantity. When the owner has to generate a bill, he/she has to enter the quantity of the product and choose the product from the list. The system will generate a bill with a random bill number and update that bill in a common database of the supermarket.